+++ app/Http/Controllers/Maintainance/CertificatesController.php
@@ -101,6 +101,12 @@ class CertificatesController extends Controller
      */
     public function destroy($id)
     {
+        if(auth()->user()->userType != 1)
+        {
+            // abort(403,'Request Unauthorized');
+            return response('Unauthorized action', 403);
+        }
+
         Certificate::destroy($id);
 
         return  [
