+++ app/Http/Controllers/Maintainance/DepartmentsController.php
@@ -100,6 +100,12 @@ class DepartmentsController extends Controller
      */
     public function destroy($id)
     {
+        if(auth()->user()->userType != 1)
+        {
+            // abort(403,'Request Unauthorized');
+            return response('Unauthorized action', 403);
+        }
+
         Department::destroy($id);
 
         return  [
