+++ app/Http/Controllers/Maintainance/IndustriesController.php
@@ -106,6 +106,12 @@ class IndustriesController extends Controller
      */
     public function destroy($id)
     {
+        if(auth()->user()->userType != 1)
+        {
+            // abort(403,'Request Unauthorized');
+            return response('Unauthorized action', 403);
+        }
+
         Industry::destroy($id);
         
         return  [
